+++ src/Functions/clear.js
@@ -1,0 +1,17 @@
+const fs = require('fs');
+
+/**
+ * Core clear function
+ * @param {*} stream 
+ */
+module.exports = function clear(stream) {
+	try {
+		if (stream === process.stdout) {
+			process.stdout.cursorTo(0,0);
+			process.stdout.clearLine(0);
+			process.stdout.clearScreenDown();
+		}
+	} catch (error) {
+		throw Error;
+	}
+}