+++ src/index.js
@@ -11,5 +11,6 @@ Object.defineProperty(exports, 'Transports', {
 	enumerable: true,
 	get() {
 		return require('./Transports/index.js');
-	}
+	},
+	writable: true
 });